# chapter-7/api-reservation/src/main/resources/openapi/api.yaml
{"openapi":"3.0.1","info":{"title":"API Reservations","description":"API for managing reservations","contact":{"name":"Andres Sacco","url":"https://example.com","email":"sacco.andres@gmail.com"},"license":{"name":"Apache 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0"},"version":"1.0"},"servers":[{"url":"http://localhost:8090/api/flights/reservation","description":"Generated server url"}],"tags":[{"name":"Reservation","description":"Operations about the reservations"}],"paths":{"/{id}":{"get":{"tags":["Reservation"],"summary":"Get a reservation by ID","description":"Fetches a reservation by its unique ID","operationId":"getReservationById","parameters":[{"name":"id","in":"path","description":"ID of the reservation to fetch","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"429":{"description":"Rate limit exceeded","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDTO"}}}},"404":{"description":"Reservation not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDTO"}}}},"200":{"description":"Reservation found successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationDTO"},"examples":{"Successful response":{"description":"Successful response","value":{"id":1,"itineraryId":"2","searchId":"2","passengers":[{"firstName":"Andres","lastName":"Sacco","documentNumber":"987654321","documentType":"DNI","nationality":"AR","birthday":"1985-01-01"}],"contact":{"telephoneNumber":"54911111111","email":"sacco.andres@gmail.com"}}}}}}}}},"put":{"tags":["Reservation"],"summary":"Update a reservation","description":"Updates the details of an existing reservation by ID","operationId":"update","parameters":[{"name":"id","in":"path","description":"ID of the reservation to update","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"requestBody":{"description":"Updated reservation details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationDTO"}}},"required":true},"responses":{"404":{"description":"Reservation not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDTO"}}}},"200":{"description":"Reservation updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationDTO"}}}},"400":{"description":"Invalid reservation data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDTO"}}}}}},"delete":{"tags":["Reservation"],"summary":"Delete a reservation","description":"Deletes a reservation by ID","operationId":"delete","parameters":[{"name":"id","in":"path","description":"ID of the reservation to delete","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"404":{"description":"Reservation not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDTO"}}}},"204":{"description":"Reservation deleted successfully","content":{"application/json":{}}}}}},"/":{"get":{"tags":["Reservation"],"summary":"Get all reservations","description":"Retrieves a list of all reservations","operationId":"getReservations","responses":{"200":{"description":"Reservations retrieved successfully","content":{"application/json":{"schema":{"type":"string"}}}}}},"post":{"tags":["Reservation"],"summary":"Save a new reservation","description":"Creates a new reservation","operationId":"save","requestBody":{"description":"Details of the reservation to create","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationDTO"},"examples":{"Valid reservation request":{"description":"Valid reservation request","value":{"itineraryId":"2","searchId":"2","passengers":[{"firstName":"Andres","lastName":"Sacco","documentNumber":"987654321","documentType":"DNI","nationality":"AR","birthday":"1985-01-01"}],"contact":{"telephoneNumber":"54911111111","email":"sacco.andres@gmail.com"}}},"Invalid reservation request":{"description":"Invalid reservation request","value":{"itineraryId":"3","searchId":"4","passengers":[{"firstName":"Maria","lastName":"Gonzalez","documentNumber":"123456789","documentType":"PASSPORT","nationality":"USS","birthday":"1990-05-20"}],"contact":{"telephoneNumber":"1234567890","email":"maria.gonzalez@example.com"}}}}}},"required":true},"responses":{"201":{"description":"Reservation created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReservationDTO"},"examples":{"Successful response":{"description":"Successful response","value":{"id":6,"itineraryId":"2","searchId":"2","passengers":[{"firstName":"Andres","lastName":"Sacco","documentNumber":"987654321","documentType":"DNI","nationality":"AR","birthday":"1985-01-01"}],"contact":{"telephoneNumber":"54911111111","email":"sacco.andres@gmail.com"}}}}}}},"400":{"description":"Invalid reservation data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDTO"}}}}}}}},"components":{"schemas":{"ContactDTO":{"required":["email","telephoneNumber"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"telephoneNumber":{"type":"string"},"email":{"type":"string"}}},"PassengerDTO":{"required":["birthday","documentNumber","documentType","firstName","lastName"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"maxLength":20,"minLength":2,"type":"string"},"lastName":{"maxLength":20,"minLength":2,"type":"string"},"documentNumber":{"maxLength":10,"minLength":2,"type":"string"},"documentType":{"maxLength":10,"minLength":3,"type":"string"},"nationality":{"type":"string"},"birthday":{"type":"string","format":"date"}}},"ReservationDTO":{"required":["contact","itineraryId","passengers","searchId"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"itineraryId":{"type":"string"},"searchId":{"type":"string"},"passengers":{"type":"array","items":{"$ref":"#/components/schemas/PassengerDTO"}},"contact":{"$ref":"#/components/schemas/ContactDTO"}}},"ErrorDTO":{"type":"object","properties":{"description":{"type":"string"},"reasons":{"type":"array","items":{"type":"string"}}}}}}}